Added feature to delay iterator's endpoint requests (including generator's construct query requests)
