+++ static/example/config.yml
@@ -14,7 +14,7 @@ stages:
     iterator:
       query: file://static/example/iterator-stage-1.rq
       endpoint: https://api.triplydb.com/datasets/Triply/iris/services/demo-service/sparql
-      delay: "30 seconds"
+      delay: "50ms"
     generator: 
       # First generator
       -  query: file://static/example/generator-stage-1-1.rq
